Rename skin -> base_skin
