+++ droppable/case.py
@@ -1,6 +1,6 @@
 from constants import Constants
 from droppable.droppable import Droppable
-from skin import Skin
+from base_skin import Skin
 from webreader import get_case_urls, get_skin_links, read_page
